feat: handel delete all cart item and change UI

+++ src/pages/cart/components/deleteModal/index.tsx
@@ -1,33 +1,48 @@
 import { useDispatch } from "react-redux";
-import { removeFromCart } from "../../../../store/cartSlice";
+import { clearCart, removeFromCart } from "../../../../store/cartSlice";
+import { persistor } from "../../../../store/store";
 
 export default function CartDeleteModal({
   productId,
   setOpen,
+  deleteAll,
+  setDeleteAll,
 }: {
   productId: string;
+  deleteAll: boolean;
   setOpen: (open: boolean) => void;
+  setDeleteAll: (open: boolean) => void;
 }) {
   const dispatch = useDispatch();
+
   function handelDelete() {
-    dispatch(removeFromCart(productId));
+    if (deleteAll) {
+      persistor.purge();
+      dispatch(clearCart());
+      setDeleteAll(false);
+    } else {
+      dispatch(removeFromCart(productId));
+    }
     setOpen(false);
   }
+
   return (
-    <div className="w-screen h-screen absolute -top-20 left-0 bg-gray-6 bg-opacity-80 flex justify-center items-center">
-      <div className="h-[200px] bg-[#fff] rounded-md flex flex-col justify-evenly items-center p-5 gap-3">
-        <p className="text-lg font-medium">
-          آیا میخواهید این محصول را حذف کنید؟
+    <div className="w-screen h-screen absolute -top-20 left-0 bg-black bg-opacity-70 flex justify-center items-center transition-opacity duration-300 ease-in-out p-7">
+      <div className="h-[200px] bg-gradient-to-r from-gray-100 to-gray-50 rounded-lg shadow-xl transform scale-105 flex flex-col justify-evenly items-center p-8 gap-5 animate-fadeIn">
+        <p className="text-xl font-semibold text-gray-800 text-center">
+          {deleteAll
+            ? "از حذف همه محصولات اطمینان دارید؟"
+            : "از حذف این محصول اطمینان دارید؟"}
         </p>
         <div className="flex justify-between gap-5 w-full">
           <button
-            className="bg-error-light w-1/2 text-[#fff] px-5 py-2 rounded-md"
+            className="w-1/2 bg-error-light text-[#fff] px-6 py-3 rounded-lg shadow-md hover:bg-error hover:scale-105 transition-all duration-200 ease-in-out"
             onClick={() => setOpen(false)}
           >
             خیر
           </button>
           <button
-            className="bg-primary w-1/2 text-[#fff] px-5 py-2 rounded-md"
+            className="w-1/2 bg-tint-3 text-[#fff] px-6 py-3 rounded-lg shadow-md hover:bg-tint-4 hover:scale-105 transition-all duration-200 ease-in-out"
             onClick={handelDelete}
           >
             بله
